填了一堆 waring 的坑

+++ MoblieCQUPT_iOS/ClassWidget/TodayViewController.m
@@ -87,7 +87,7 @@ fprintf(stderr, "-------\n");                                               \
                                       @"begin_lesson":@"-1"}
                                       ] mutableCopy];
     }
-    NSLog(@"今日周%ld,数据:%@",today,mutableToDayClassArray);
+    NSLog(@"今日周%ld,数据:%@",(long)today,mutableToDayClassArray);
     return mutableToDayClassArray;
 }
 
@@ -121,7 +121,7 @@ fprintf(stderr, "-------\n");                                               \
 
 #pragma UI 相关
 - (void) updateView{
-    NSLog(@"更新界面:%ld",self.todayClassArray.count);
+    NSLog(@"更新界面:%ld",(unsigned long)self.todayClassArray.count);
     [self.classTableView reloadData];
     self.preferredContentSize = CGSizeMake([UIScreen mainScreen].bounds.size.width, kTableViewCellRowHeight*self.todayClassArray.count);
     self.classTableView.frame = self.view.bounds;
@@ -175,7 +175,7 @@ fprintf(stderr, "-------\n");                                               \
 
 - (NSString *)stringWithBeginLesson:(NSInteger)beginLesson
                              period:(NSInteger)time{
-    NSLog(@"%ld==%ld",beginLesson,time);
+    NSLog(@"%ld==%ld",(long)beginLesson,(long)time);
     NSString *startTimeString,*endTimeString,*string;
     NSInteger baseClassNum = 1;
     if (beginLesson == -1) {
